Change to using directories for files
Means we don't have to have .html in URLs for posts, which is--I guess
an advantage? I found something charming about the return to html, but
that's probably irrational. And this will let us put comments as a file
or files inside of posts directories so we don't rely on some kind of
name matchup to build them together.